requirements.txt
Flask==2.3.3
requests==2.31.0
python-docx==0.8.11
python-pptx==0.6.21
PyPDF2==3.0.1
Pillow==10.0.1
pytesseract==0.3.10
matplotlib==3.7.2
networkx==3.1
numpy==1.24.3
werkzeug==2.3.7